Fixing instantiation of error in success?
InternalError was not being instantiated properly. This changes the code
to instantiate UnknownError with a single argument. Also adds a small
test to verify that the exception is thrown.

This also deletes some integration tests that try to talk to S3. These
tests have never passed because they require entering real AWS
credentials and an S3 bucket.